store-api-sql: validação de estoque na exclusão da venda

+++ Módulo 3 - Bancos de Dados + Node.js/banco-de-dados-sql/store-api-sql/services/sale.service.js
@@ -39,7 +39,16 @@ async function getSale(id) {
 }
 
 async function deleteSale(id) {
-  await SaleRepository.deleteSale(id);
+  const sale = await SaleRepository.getSale(id);
+
+  if (sale) {
+    const product = await ProductRepository.getProduct(sale.product_id);
+    await SaleRepository.deleteSale(id);
+    product.stock++;
+    await ProductRepository.updateProduct(product);
+  } else {
+    throw new Error('O id da sale informado não existe.');
+  }
 }
 
 async function updateSale(sale) {
